Move Inventory and Artifact to inventory utils
