+++ common/go-utils/src/lib.rs
@@ -1,2 +1,4 @@
+use toml as _;
+
 pub mod inv;
 pub mod vrs;
