+++ common/inventory-utils/Cargo.toml
@@ -10,3 +10,4 @@ workspace = true
 semver = "1"
 serde = { version = "1", features = ["derive"] }
 thiserror = "1"
+toml = "0.8"
