+++ common/inventory-utils/src/lib.rs
@@ -1,2 +1,3 @@
 pub mod checksum;
+pub mod inv;
 pub mod vrs;
